Add entropy calculations to gas container
Add new utilities to Kinematica for quickly computing logs of factorials
and binomial functions (see previous commit). Compute entropy where
macrostate is number of balls on left and microstate is number of valid
rearrangements. Display the current entropy as well as the max possible
entropy for this number of balls.

Will look great once we get real time plotting in here.