Add fullscreen & resize; hide dat.gui; boost performance
[+] Fullscreen
Add a div around the container called "#fullscreen". Add fontawesome
glyphicons, and use the fullscreen glyphicon as a button for fullscreen
(currently works in webkit browsers, i.e. chrome and safari).

[+] Responsive resizing
Add responsive resizing by changing every line that mentions "width" or
"height" to "stage.getWidth()" or "stage.getHeight()" so that width and
height are always the latest values for the stage. Also add CSS for the
"#fullscreen" selector to make it the full size of the viewport (i.e.
the client's browser window). Add event handler for when the window is
resized, which resizes the stage and moves the left and right rectangles
so that they are in the correct positions. Change the size of the rectangles
when they are constructed: the rectangles are now far larger than necessary
(10x the initial size of the viewport), meaning they likely don't have to
be resized when the window changes size. This improves performance during
resizing.

[+] Hide dat.gui
The dat.gui controls are initially closed, but can be opened by pressing
the "open controls" button, which remains visible in the top right corner.
This is particularly aimed at mobile browsers with limited screen size.

[+] Improved performance
Now, when a ball is created, it is immediately cached as an image. In safari,
the performance difference is negligible (performance was already excellent),
but in Chrome, the performance went from bad to silky smooth.